SubBiome sampling (without choosing which zones are valid for centers)
Next up is to determine which parts of the map should be considered valid for Sub-Biome centers. Current plan is to use a radial constraint from the center, but nothing is confirmed yet. Some Biome centers might be lakes/mountaintops/islands/etc as a rule instead, but nothing for sure yet